Preload workspace fonts on home and customize views
